Remove cards/*/learn action
This action was meant to allow the user to learn one specific card.
Since we won't the user do that we got rid of this action.

In order to make this work for the POST method type we now provide
the id of the card that was learned via a hidden input field in the
learn form.