Ensure maximum size of card
The additional card-content-wrapper allows us to move the display: table
style from the .card to the .card-content-wrapper element. This ensures
that the width and height (that are set on .card) cannot be exceeded since
tables will always cover as much space as they need (thus ignoring any
width and height attributes).