Card/add - fix misaligned submit button
The changes in d1101ac00b30b5adb1dfa0ca13acb0782af6290c added a cancel
button on the edit page. The new flexbox styles that came with that broke
the submit button on the add page however since the margin: 0 auto; style
had been removed.

This commit fixes this.