Tests - avoid errors when encountering HTML5 tags
The XML::LibXML parser looks for a DTD at the start of the document
in order to get a definition of all tags that can be used in said
document. Since HTML5 however an external DTD is not needed anymore.
Instead you just use <!doctype html>.

Therefore XML::LibXML falls back to the default HTML4 tag names and
will crash when it encounters HTML5-specific tags like <header>.

The solution for this problem was to activate the 'recover' mode for
the HTML parser.

More info on this can be found here:

https://metacpan.org/pod/XML::LibXML::Parser#recover