only allow to learn current user's cards
Since the introduction of user-specific cards we need to know the
current user's id in order to retrieve only their cards.

Therefore we need to either

 a) pass the user's id into Model::DB's get_due_cards sub

 b) move the "get_due_cards" logic into Controller::Cards

We chose option b) since this will allow us to re-use the existing
code that retrieves the resultset that contains all user's cards.
We can then further reduce these cards to only those cards that are
due to learn.